Photos on main can be turned over (but the quality is law)

+++ app/src/main/java/com/safiya/photogallery/FeedReaderDbHelper.kt
@@ -50,8 +50,7 @@ class FeedReaderDbHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_
                 val tagsString = getString(getColumnIndexOrThrow(FeedReaderContract.PhotoEntry.COLUMN_NAME_TAGS))
                 val tags = tagsString.split(",").toTypedArray() // Преобразуем строку в массив
 
-                photos.add(Photo(id.toInt(), deviceId, createdAt, imageData, title, tags))
-            }
+                photos.add(Photo(id, deviceId, createdAt, imageData, title, tags))            }
         }
         cursor.close()
         return photos
