+++ app/src/main/java/com/safiya/photogallery/PhotoRepository.kt
@@ -39,8 +39,7 @@ class PhotoRepository(private val dbHelper: FeedReaderDbHelper) {
                 val tagsString = getString(getColumnIndexOrThrow(FeedReaderContract.PhotoEntry.COLUMN_NAME_TAGS))
                 val tags = tagsString.split(",").toTypedArray() // Преобразуем строку в массив
 
-                photos.add(Photo(id.toInt(), deviceId, createdAt, imageData, title, tags))
-            }
+                photos.add(Photo(id, deviceId, createdAt, imageData, title, tags))            }
         }
         cursor.close()
         return photos
